Moving extractor specific ruletypes to extractor
- This is to avoid mixing the direct usage of yielded chunks within the
  ruletypes which is needed for the extractor, but detracts from the
  original design of those callables where it should simply pass the
  chunks through.
- Ideally that whole underlying API should be formalized.
